+++ include/Component.h
@@ -19,7 +19,7 @@ public:
     Component(GameObject& associated);
     virtual ~Component();
 
-    void virtual Update(float dt) = 0;
+    void virtual Update(double dt) = 0;
     void virtual Render() = 0;
     bool virtual Is(std::string type) = 0;
     const std::string virtual GetType() = 0;
